#669 refactor away from jaar_config.py
