+++ src/f04_kick/kick.py
@@ -1,6 +1,6 @@
 from src.f00_instrument.file import save_file, open_file, open_json, create_path
 from src.f00_instrument.dict_toolbox import get_json_from_dict, get_dict_from_json
-from src.f01_road.jaar_config import get_init_kick_id_if_None, get_json_filename
+from src.f01_road.jaar_config import get_json_filename
 from src.f01_road.road import (
     FaceName,
     OwnerName,
@@ -20,6 +20,14 @@ from os.path import exists as os_path_exists
 
 class kick_bud_conflict_Exception(Exception):
     pass
+
+
+def init_kick_id() -> int:
+    return 0
+
+
+def get_init_kick_id_if_None(x_kick_id: int = None) -> int:
+    return init_kick_id() if x_kick_id is None else x_kick_id
 
 
 @dataclass
