+++ src/f06_listen/hubunit.py
@@ -15,10 +15,7 @@ from src.f01_road.jaar_config import (
     grades_folder,
     get_rootpart_of_keep_dir,
     treasury_filename,
-    get_kicks_folder,
-    get_init_kick_id_if_None,
     get_json_filename,
-    init_kick_id,
 )
 from src.f01_road.finance import (
     default_fund_coin_if_None,
@@ -56,7 +53,13 @@ from src.f04_kick.atom import (
     get_from_json as budatom_get_from_json,
     modify_bud_with_budatom,
 )
-from src.f04_kick.kick import KickUnit, kickunit_shop, create_kickunit_from_files
+from src.f04_kick.kick import (
+    get_init_kick_id_if_None,
+    init_kick_id,
+    KickUnit,
+    kickunit_shop,
+    create_kickunit_from_files,
+)
 from src.f06_listen.basis_buds import get_default_plan
 from src.f06_listen.hub_path import create_gut_path, create_plan_path
 from src.f06_listen.hub_tool import (
@@ -144,7 +147,7 @@ class HubUnit:
         self._owner_dir = create_path(self._owners_dir, self.owner_name)
         self._keeps_dir = create_path(self._owner_dir, "keeps")
         self._atoms_dir = create_path(self._owner_dir, "atoms")
-        self._kicks_dir = create_path(self._owner_dir, get_kicks_folder())
+        self._kicks_dir = create_path(self._owner_dir, "kicks")
         self._deals_dir = create_path(self._owner_dir, "deals")
 
     def default_gut_bud(self) -> BudUnit:
